fix: maybe_coerce_error - change expected input and outputs
This method is only used to decode a failed simulation, which means it should get a SimulationError as the input error type. It returns one of two types of recoverable errors - TransactionError and StorageError.

Did not yet introduce a NoneRecoverableError type as this has not yet appeared in any method (RuntimeError).