fix: (VMPoolState) check for hard limit in get_amount_out
implement get_amount_out_sell_limit to test this.

Note: The actual values of the error cannot be tested without deriving PartialEq - which is a pain since some lower-level errors don't implement this, so we'd have to implement it manually. I've manually checked that the values made sense though... it would be nice to improve this test one day.