feat(simulation): try to bruteforce token slots when they are unknown
Add a function that will try to bruteforce balance and allowance slots for a given token. This function will iterate over the mapping "base_slot" and try to set a marker value to a particular key and read it. If the value is correctly retrieved, it means the "base_slot" is correct.