feat: add `VMPoolStateBuilder` and refactor TychoSimulationContract to improve engine handling
This commit aims to abstract away a lot of complexity that is in `VMPoolState` using a builder pattern.

This pattern helps us separating the initialisation and the execution of `VMPoolState`, making it more straightforward and simpler. For example many field where given to VMPoolState just for initialisation and never used after (`engine`, `trace`, ect..). Some where also `Option` but actually required by the struct at runtime (like `adapter_contract`.

The builder also allow any field to be directly given, for example if someone wants to manually create the `AdapterContract`.